drm/msm/dp: Set struct drm_bridge::ddc to link ddc to DP device
The ddc is created for the drm_dp_aux device, but it isn't associated
with the bridge. This makes finding the DDC for the DP card in sysfs
(card0-DP-1 for example) hard because there isn't a link between the i2c
device created underneath the displayport controller device and the DRM
card device. Set the struct drm_bridge ddc member to create this
linkage in sysfs. This helps userspace understand the relationship
between the i2c device and the DP card.