platform/chrome: cros_ec_typec: Allow DP configure to work
The DP altmode driver fails the configure stage because the status VDO
that is spoofed in cros_typec_enable_dp() is missing a couple flags. Add
them so that the configure succeeds. This has the nice side effect of
properly reflecting the pin assignment and configuration of the DP
altmode in sysfs.

Signed-off-by: Stephen Boyd <swboyd@chromium.org>