drm/connector: Consider parent in drm_connector_find_by_fwnode()
Sometimes the fwnode associated with the drm_connector is the parent of
the fwnode that is being searched. For example, USB Type-C DP altmode
will register a drm_bridge for the TCPM, which is the parent of the
usb-c-connector node. The drm_bridge won't extend to the usb-c-connector
node itself because the DP signal is muxed to one of many USB type-c
ports.